Simplify advanced example
