+++ package.json
@@ -20,7 +20,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "svix": "^0.78.0",
-    "svix-react": "^1.4.2"
+    "svix-react": "^1.7.0"
   },
   "devDependencies": {
     "@types/node": "18.13.0",
